ARM: dts: tegra20-glide: USB OTG bindings
OTG here is provided by ChipIdea controller, which uses OTG FSM to
power on the connected device. However, OTG specification says that
the device needs to receive power in 100ms, but this is not a strict
requirement. We need to disable the charger first and then power on VBUS,
which is slow process so we set a delay for the regulator to workaround
this.

Signed-off-by: Sergey Larin <cerg2010cerg2010@mail.ru>