HACK: ARM: seccomp: Enforce whitelisting of clock_gettime64 syscall
After updating to Ubuntu 20.04, I found that all chromium-based software
got broken. This happens because Ubuntu 20.04 updated GLIBC to 2.31,
which now uses clock_gettime64 by default, but chromium doesn't whitelist
this syscall for seccomp. The clock_gettime64 is a VDSO function on ARM,
and thus, there is no good reason to fail software that is not ready
for a newer GLIBC versions. Falkon and Akregator are now working on
Ubuntu 20.04 without a need to disable seccomp by software, which isn't
always possible (like in a case of Akregator).

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>