partitions/tegra: Support gpt_sector= command line option
The gpt_sector=<sector> causes the GPT partition search to look at the
specified sector for a valid GPT header if the GPT is not found at the
beginning or the end of block device.  This is needed by Tegra-based
devices that have TegraPT and GPT placed at inaccessible by kernel
locations.  The GPT entry duplicates TegraPT partitions.

This patch is based on the original work done by Colin Cross for the
downstream Android kernel.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>